[BE] ADD : 내 진행 크루 보기 추가

+++ BE/gumid101/src/main/java/com/ssafy/gumid101/entity/UserCrewJoinEntity.java
@@ -28,7 +28,7 @@ public class UserCrewJoinEntity {
 	@Id
 	@Column(name="crew_user_seq")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int crewUserSeq;
+	private Long crewUserSeq;
 
 	@ManyToOne
 	@JoinColumn(name="user_seq")
